pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plex-recommender"
version = "0.1.0"
description = "AI-powered recommendation engine for Plex using Ollama"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "typer[all]>=0.9.0",
    "plexapi>=4.15.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "apscheduler>=3.10.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "lancedb>=0.17.0",
    "pyarrow>=15.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
plex-rec = "plex_recommender.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",  # Line too long - handled by formatter where possible
]

[tool.ruff.lint.per-file-ignores]
"plex_recommender/recommend/engine.py" = ["N802"]  # Allow SYSTEM_PROMPT property name
